+++ CHANGES.md
@@ -7,6 +7,7 @@ NEXT VERSION
 
 **Features:**
 
+- Improved readability of asynchronous tests: `assertWithTimeout(1, thatEventually(var), is(@10));`
 - Added assertions designed for use in standalone functions: `assertThatC` and `assertWithTimeoutC`.
 - Added ability to add custom test failure handlers. See HCTestFailureHandlerChain.
 - Added optional test failure handler for Cedar. See HCCedarTestFailureHandler and add to HCTestFailureHandlerChain. Use the new standalone assertions `assertThatC` and `assertWithTimeoutC`.